requirements.txt
streamlit==1.32.2
pandas==2.2.1
langchain==0.1.13
langchain-google-genai==1.0.1
langchain-community==0.0.29
faiss-cpu==1.8.0
python-dotenv==1.0.1
plotly==5.20.0
tiktoken==0.6.0
